requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
seaborn==0.12.2
scikit-learn==1.2.2
streamlit==1.20.0
tensorflow==2.12.0
opencv-python==4.7.0.72